Move override file read
