+++ src/swell/deployment/create_experiment.py
@@ -31,19 +31,6 @@ from swell.utilities.check_da_params import check_da_params
 # --------------------------------------------------------------------------------------------------
 
 
-def read_override_file(override_path: str | None) -> dict:
-
-    yaml = YAML(typ='safe')
-
-    if override_path is None:
-        return {}
-    else:
-        with open(override_path, 'r') as f:
-            return yaml.load(f)
-
-# --------------------------------------------------------------------------------------------------
-
-
 def clone_config(
     configuration: str,
     experiment_id: str,
@@ -233,7 +220,7 @@ def create_experiment_directory(
     suite_config: str,
     method: str,
     platform: str,
-    override: str | dict,
+    override: dict,
     advanced: bool,
     slurm: str | None,
     skip_r2d2: bool
@@ -247,26 +234,17 @@ def create_experiment_directory(
     # ---------------
     logger = get_logger('SwellCreateExperiment')
 
-    # Read override file
-    # ------------------
-    if isinstance(override, str):
-        override_dict = read_override_file(override)
-    elif override is None:
-        override_dict = {}
-    else:
-        override_dict = override
-
     # Specify whether to skip registering and storing in R2D2
     # -------------------------------------------------------
     if skip_r2d2:
 
         # Only override this if it is true, otherwise let the suite decide
-        override_dict['skip_r2d2'] = skip_r2d2
+        override['skip_r2d2'] = skip_r2d2
 
     # Call the experiment config and suite generation
     # ------------------------------------------------
     experiment_dict_str = prepare_config(suite, suite_config, method, platform,
-                                         override_dict, advanced, slurm)
+                                         override, advanced, slurm)
 
     # Load the string using yaml
     # --------------------------
