+++ src/swell/utilities/suite_utils.py
@@ -11,6 +11,7 @@
 import glob
 import os
 import importlib
+from ruamel.yaml import YAML
 
 from swell.swell_path import get_swell_path
 
@@ -75,5 +76,16 @@ def get_suite_tests() -> list:
     # Return list of valid task choices
     return suite_tests
 
+
+# --------------------------------------------------------------------------------------------------
+
+def read_override_file(override_path: str | None) -> dict:
+
+    if override_path is None:
+        return {}
+    else:
+        yaml = YAML(typ='safe')
+        with open(override_path, 'r') as f:
+            return yaml.load(f)
 
 # --------------------------------------------------------------------------------------------------
